requirements.txt
amqp==2.6.1
asgiref==3.2.7
astroid==2.4.2
billiard==3.6.3.0
boto3==1.14.0
botocore==1.17.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.7
certifi==2020.4.5.2
cffi==1.14.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.3.0
dateutils==0.6.8
defusedxml==0.6.0
dicttoxml==1.7.4
Django==3.0.6
django-cleanup==5.0.0
django-filter==2.3.0
django-rest-framework-braces==0.3.4
django-seed==0.2.2
djangorestframework==3.11.0
docutils==0.15.2
drf-yasg==1.17.1
Faker==4.1.1
html5lib==1.1
idna==2.9
inflection==0.5.0
install==1.3.3
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==4.6.11
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.4
Pillow==7.2.0
psycopg2-binary==2.8.5
pycparser==2.20
pylint==2.5.3
pyparsing==2.4.7
Pyphen==0.9.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
stripe==2.49.0
text-unidecode==1.3
tinycss2==1.0.2
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
wrapt==1.12.1
rawlerdetect==0.1.4